sarah-naqvi/lab-5/pyproject.toml
[project]
name = "lab-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "symusic==0.5.9",
]
